RMagick - Image padding script:
Was supplied images at all different resolutions and aspect ratios.
This posed an issue for the video wall we were displaying this content to.

The dimensions of the video wall are 11520x4320 pixels, giving an aspect ratio of approximately 2.67.

In order for the images to not be cut off or to actually display appropriately, black (white-space) needed to be added to the image.
A black image was created that matched the height of the selected image and the width was then calculated to match the 2.67 aspect ratio.
Now that two images are available, the image to bne displayed was then overlayed on the new black image, essentially creating a combined composite image.
A gravity configuration was applied for this composite function which allowed the image to actually be centered all the time.

This was then applied to 96 other images where the file names were stored in an array.